+++ ui/src/components/Graphs/NoData.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 export default () => (
-  <h4 style={{ textAlign: 'center', marginTop: 20 }}>
+  <h4 style={{ textAlign: 'center', marginTop: 20, width: '100%' }}>
     No data in this time range.
   </h4>
 );