+++ ui/src/containers/DashboardGrid/index.js
@@ -76,7 +76,7 @@ const getLayout = widgets =>
       y: widget.get('y', 0),
       w: widget.get('w', 4),
       h: widget.get('h', 4),
-      minW: 2,
+      minW: 3,
       minH: 4
     };
   });